Fix lint error

+++ server/app/controllers/me-controller.js
@@ -126,7 +126,7 @@ async function addBarmanService(req, res) {
     const { error } = schema.validate(req.body);
     if (error) throw createUserError('BadRequest', error.message);
 
-    if(!req.barman.active) throw createUserError('NotActive', 'Provided barman is not an active barman');
+    if (!req.barman.active) throw createUserError('NotActive', 'Provided barman is not an active barman');
     const servicesId = req.body;
 
     await barmanService.createServiceBarman(req.barman.id, servicesId);
@@ -146,6 +146,7 @@ async function removeBarmanService(req, res) {
     const { error } = schema.validate(req.body);
     if (error) throw createUserError('BadRequest', error.message);
 
+    if (!req.barman.active) throw createUserError('NotActive', 'Provided barman is not an active barman');
     const servicesId = req.body;
 
     await barmanService.deleteServiceBarman(req.barman.id, servicesId);
